requirements.txt
GeoIP
pytelegrambotapi
requests
pillow
netaddr
requests
BeautifulSoup4
selenium
pymongo
urllib3
zmq
jsoncomment
rq
pyyaml